ch2 refactor, updated launch.json
