+++ .vscode/launch.json
@@ -5,7 +5,7 @@
     "version": "0.2.0",
     "configurations": [
         {
-            "name": "Object",
+            "name": "Chapter 2",
             "type": "go",
             "request": "launch",
             "mode": "debug",
